requirement.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
bzip2==1.0.8
cachetools==5.3.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
docx==0.2.4
emoji==2.9.0
exceptiongroup==1.2.0
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
gitpython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langdetect==1.0.9
langsmith==0.0.80
libffi==3.4.4
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
markupsafe==2.1.3
marshmallow==3.20.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.7.2
openssl==3.0.12
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pinecone-client==2.2.4
pip==23.3.1
protobuf==4.25.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic-core==2.14.6
pydeck==0.8.1b0
pygments==2.17.2
pymupdf==1.23.14
pymupdfb==1.23.9
python==3.10.13
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-magic==0.4.27
pyyaml==6.0.1
rapidfuzz==3.6.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.25
sqlite==3.41.2
streamlit==1.30.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tk==8.6.12
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.12.0
unstructured-client==0.15.2
urllib3==2.1.0
validators==0.22.0
vc==14.2
watchdog==3.0.0
wheel==0.41.2
win32-setctime==1.1.0
wrapt==1.16.0
xz==5.4.5
yarl==1.9.4
zipp==3.17.0
zlib==1.2.13
